+++ packages/cms/cmscore/src/index.ts
@@ -1,3 +1,5 @@
+export { version } from '../package.json';
+
 export { CMSList } from './CMSList';
 export { CMSItem } from './CMSItem';