+++ packages/cms/cmssort/src/buttons.ts
@@ -1,5 +1,7 @@
-import { sortListItems } from './sort';
+import { checkCMSCoreVersion } from '$cms/utils/versioning';
+import { normalizePropKey } from '$cms/utils/props';
 import { ATTRIBUTES } from './constants';
+import { sortListItems } from './sort';
 
 import type { ButtonState, ButtonsState, CSSClasses, SortingDirection, SortItemsCallback } from './types';
 import type { CMSItem, CMSList } from '$cms/cmscore/src';
@@ -84,7 +86,10 @@ export const initButtons = (
  * @param globalCSSClasses The state CSS classes (`asc` and `desc`) globally defined on the list.
  */
 const prepareButton = (button: HTMLElement, buttonsState: ButtonsState, globalCSSClasses: CSSClasses) => {
-  const sortKey = button.getAttribute(ATTRIBUTES.field.key);
+  // `cmscore v1.2.0` implements propKeys normalization.
+  // TODO: Make this a default after 24th November.
+  let sortKey: string | null | undefined = button.getAttribute(ATTRIBUTES.field.key);
+  if (checkCMSCoreVersion('>=', '1.2.0')) sortKey = normalizePropKey(sortKey);
   if (!sortKey) return;
 
   const reverse = button.getAttribute(reverseKey) === reverseValues.true;