+++ packages/cms/cmsload/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `cmsload` Changelog
 
+## [v1.2.0] 17th November 2021
+
+- Added support for `cmscore` versioning.
+
 ## [v1.1.5] 16th November 2021
 
 - Updated `cmscore` import.